Enable Serialization/Deserialization round-trip for Domain objects
The custom validator on the 'created' and 'expires' field allowed for
either a datetime object directly, or Namecheap's 'MM/DD/YYYY' string
format. But if you serialize the object (ie, model_dump_json()) and then
try to deserialize it using the Domain model, it would fail because
those fields would have strings of the format 'YYYY-MM-DDThh-mm-ss'
which wouldn't validate.

This adds a quick if/else to see if there is a 'T' in the datetimestamp
string, if so it uses datetime.fromisoformat(v), rather than the
datetime.strptime(v, "%m/%d/%Y').

This change allows a Domain object to be serialized to JSON, and
reserialized back into a Domain object.